New: Use alternative names to detect IGDB records when downloading metadata #391
New: Use additional fields when downloading metadata from IGDB #571
Fix: Reboots have cover art from original game #404
