+++ source/Playnite/Providers/Uplay/IUplayLibrary.cs
@@ -1,4 +1,4 @@
-﻿using Playnite.Models;
+﻿using Playnite.Metadata;
 using Playnite.SDK.Models;
 using System;
 using System.Collections.Generic;
