+++ source/PlayniteServices/PlayniteServices.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2035
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlayniteServices", "PlayniteServices.csproj", "{5C36DDAA-F490-451C-9EF8-F949683F3821}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5C36DDAA-F490-451C-9EF8-F949683F3821}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C36DDAA-F490-451C-9EF8-F949683F3821}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C36DDAA-F490-451C-9EF8-F949683F3821}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C36DDAA-F490-451C-9EF8-F949683F3821}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EFEC98D8-5168-4573-9B5A-037A1D674EFF}
+	EndGlobalSection
+EndGlobal
