+++ source/PlayniteUI/ViewModels/GameEditViewModel.cs
@@ -19,6 +19,7 @@ using Playnite.Providers.BattleNet;
 using Playnite.SDK;
 using Playnite.SDK.Models;
 using Playnite.Web;
+using Playnite.Metadata;
 
 namespace PlayniteUI.ViewModels
 {
@@ -2365,17 +2366,17 @@ namespace PlayniteUI.ViewModels
                     
                     if (metadata.Image != null)
                     {
-                        var path = Path.Combine(Paths.TempPath, metadata.Image.Name);
+                        var path = Path.Combine(Paths.TempPath, metadata.Image.FileName);
                         FileSystem.PrepareSaveFile(path);
-                        File.WriteAllBytes(path, metadata.Image.Data);
+                        File.WriteAllBytes(path, metadata.Image.Content);
                         tempGame.Image = path;
                     }
 
                     if (metadata.Icon != null)
                     {
-                        var path = Path.Combine(Paths.TempPath, metadata.Icon.Name);
+                        var path = Path.Combine(Paths.TempPath, metadata.Icon.FileName);
                         FileSystem.PrepareSaveFile(path);
-                        File.WriteAllBytes(path, metadata.Icon.Data);
+                        File.WriteAllBytes(path, metadata.Icon.Content);
                         tempGame.Icon = path;
                     }
