+++ source/PlayniteUI/Windows/MetadataDownloadWindow.xaml
@@ -10,7 +10,7 @@
         xmlns:pui="clr-namespace:PlayniteUI"
         xmlns:controls="clr-namespace:PlayniteUI.Controls"
         xmlns:pvm="clr-namespace:PlayniteUI.ViewModels"
-        xmlns:pmp="clr-namespace:Playnite.MetaProviders;assembly=Playnite"
+        xmlns:pmp="clr-namespace:Playnite.Metadata;assembly=Playnite"
         mc:Ignorable="d"
         d:DesignStyle="{StaticResource WindowDesignStyle}"
         Style="{DynamicResource StandardWindowStyle}"
