+++ source/PlayniteUITests/ViewModels/GameEditViewModelTests.cs
@@ -28,17 +28,17 @@ namespace PlayniteUITests.ViewModels
                 var origImage = PlayniteUITests.CreateFakeFile();
                 var origBackground = PlayniteUITests.CreateFakeFile();
 
-                db.AddFile(origIcon.Name, origIcon.Name, origIcon.Data);
-                db.AddFile(origImage.Name, origImage.Name, origImage.Data);
-                db.AddFile(origBackground.Name, origBackground.Name, origBackground.Data);
+                db.AddFile(origIcon.FileName, origIcon.FileName, origIcon.Content);
+                db.AddFile(origImage.FileName, origImage.FileName, origImage.Content);
+                db.AddFile(origBackground.FileName, origBackground.FileName, origBackground.Content);
 
                 var game = new Game()
                 {
                     ProviderId = "testid",
                     Name = "Test Game",
-                    Image = origImage.Name,
-                    Icon = origIcon.Name,
-                    BackgroundImage = origBackground.Name
+                    Image = origImage.FileName,
+                    Icon = origIcon.FileName,
+                    BackgroundImage = origBackground.FileName
                 };
 
                 db.AddGame(game);
@@ -49,34 +49,34 @@ namespace PlayniteUITests.ViewModels
                 
                 // Images are replaced
                 var model = new GameEditViewModel(game, db, new MockWindowFactory(), new MockDialogsFactory(), new MockResourceProvider());
-                model.EditingGame.Icon = newIcon.Path;
-                model.EditingGame.Image = newImage.Path;
-                model.EditingGame.BackgroundImage = newBackground.Path;
+                model.EditingGame.Icon = newIcon.FileId;
+                model.EditingGame.Image = newImage.FileId;
+                model.EditingGame.BackgroundImage = newBackground.FileId;
                 model.ConfirmDialog();
 
-                Assert.AreNotEqual(game.Icon, origIcon.Name);
-                Assert.AreNotEqual(game.Image, origImage.Name);
-                Assert.AreNotEqual(game.BackgroundImage, origBackground.Name);
-                Assert.AreNotEqual(game.Icon, newIcon.Path);
-                Assert.AreNotEqual(game.Image, newImage.Path);
-                Assert.AreNotEqual(game.BackgroundImage, newBackground.Path);
+                Assert.AreNotEqual(game.Icon, origIcon.FileName);
+                Assert.AreNotEqual(game.Image, origImage.FileName);
+                Assert.AreNotEqual(game.BackgroundImage, origBackground.FileName);
+                Assert.AreNotEqual(game.Icon, newIcon.FileId);
+                Assert.AreNotEqual(game.Image, newImage.FileId);
+                Assert.AreNotEqual(game.BackgroundImage, newBackground.FileId);
 
                 var dbFiles = db.Database.FileStorage.FindAll().ToList();
                 Assert.AreEqual(3, dbFiles.Count());
 
                 using (var str = db.GetFileStream(game.Icon))
                 {
-                    CollectionAssert.AreEqual(newIcon.Data, str.ToArray());
+                    CollectionAssert.AreEqual(newIcon.Content, str.ToArray());
                 }
 
                 using (var str = db.GetFileStream(game.Image))
                 {
-                    CollectionAssert.AreEqual(newImage.Data, str.ToArray());
+                    CollectionAssert.AreEqual(newImage.Content, str.ToArray());
                 }
 
                 using (var str = db.GetFileStream(game.BackgroundImage))
                 {
-                    CollectionAssert.AreEqual(newBackground.Data, str.ToArray());
+                    CollectionAssert.AreEqual(newBackground.Content, str.ToArray());
                 }
 
                 // Duplicates are kept and not replaced
@@ -85,9 +85,9 @@ namespace PlayniteUITests.ViewModels
                 var currentBack = game.BackgroundImage;
 
                 model = new GameEditViewModel(game, db, new MockWindowFactory(), new MockDialogsFactory(), new MockResourceProvider());
-                model.EditingGame.Icon = newIcon.Path;
-                model.EditingGame.Image = newImage.Path;
-                model.EditingGame.BackgroundImage = newBackground.Path;
+                model.EditingGame.Icon = newIcon.FileId;
+                model.EditingGame.Image = newImage.FileId;
+                model.EditingGame.BackgroundImage = newBackground.FileId;
                 model.ConfirmDialog();
 
                 dbFiles = db.Database.FileStorage.FindAll().ToList();
@@ -110,31 +110,31 @@ namespace PlayniteUITests.ViewModels
                 var origIcon = PlayniteUITests.CreateFakeFile();
                 var origImage = PlayniteUITests.CreateFakeFile();
                 var origBackground = PlayniteUITests.CreateFakeFile();
-                db.AddFile(origIcon.Name, origIcon.Name, origIcon.Data);
-                db.AddFile(origImage.Name, origImage.Name, origImage.Data);
-                db.AddFile(origBackground.Name, origBackground.Name, origBackground.Data);
+                db.AddFile(origIcon.FileName, origIcon.FileName, origIcon.Content);
+                db.AddFile(origImage.FileName, origImage.FileName, origImage.Content);
+                db.AddFile(origBackground.FileName, origBackground.FileName, origBackground.Content);
                 db.AddGame(new Game()
                 {
                     ProviderId = "testid",
                     Name = "Test Game",
-                    Image = origImage.Name,
-                    Icon = origIcon.Name,
-                    BackgroundImage = origBackground.Name
+                    Image = origImage.FileName,
+                    Icon = origIcon.FileName,
+                    BackgroundImage = origBackground.FileName
                 });
 
                 origIcon = PlayniteUITests.CreateFakeFile();
                 origImage = PlayniteUITests.CreateFakeFile();
                 origBackground = PlayniteUITests.CreateFakeFile();
-                db.AddFile(origIcon.Name, origIcon.Name, origIcon.Data);
-                db.AddFile(origImage.Name, origImage.Name, origImage.Data);
-                db.AddFile(origBackground.Name, origBackground.Name, origBackground.Data);
+                db.AddFile(origIcon.FileName, origIcon.FileName, origIcon.Content);
+                db.AddFile(origImage.FileName, origImage.FileName, origImage.Content);
+                db.AddFile(origBackground.FileName, origBackground.FileName, origBackground.Content);
                 db.AddGame(new Game()
                 {
                     ProviderId = "testid2",
                     Name = "Test Game 2",
-                    Image = origImage.Name,
-                    Icon = origIcon.Name,
-                    BackgroundImage = origBackground.Name
+                    Image = origImage.FileName,
+                    Icon = origIcon.FileName,
+                    BackgroundImage = origBackground.FileName
                 });
 
                 Assert.AreEqual(6, db.Database.FileStorage.FindAll().ToList().Count());
@@ -146,9 +146,9 @@ namespace PlayniteUITests.ViewModels
                 // Replaces all images for all games
                 var games = db.GamesCollection.FindAll().ToList();
                 var model = new GameEditViewModel(games, db, new MockWindowFactory(), new MockDialogsFactory(), new MockResourceProvider());
-                model.EditingGame.Icon = newIcon.Path;
-                model.EditingGame.Image = newImage.Path;
-                model.EditingGame.BackgroundImage = newBackground.Path;
+                model.EditingGame.Icon = newIcon.FileId;
+                model.EditingGame.Image = newImage.FileId;
+                model.EditingGame.BackgroundImage = newBackground.FileId;
                 model.ConfirmDialog();
 
                 var dbFiles = db.Database.FileStorage.FindAll().ToList();
@@ -170,21 +170,21 @@ namespace PlayniteUITests.ViewModels
                 newIcon = PlayniteUITests.CreateFakeFile();
                 newImage = PlayniteUITests.CreateFakeFile();
                 newBackground = PlayniteUITests.CreateFakeFile();
-                db.AddFile(newIcon.Name, newIcon.Name, newIcon.Data);
-                db.AddFile(newImage.Name, newImage.Name, newImage.Data);
-                db.AddFile(newBackground.Name, newBackground.Name, newBackground.Data);
-                games[0].Icon = newIcon.Name;
-                games[0].Image = newImage.Name;
-                games[0].BackgroundImage = newBackground.Name;
+                db.AddFile(newIcon.FileName, newIcon.FileName, newIcon.Content);
+                db.AddFile(newImage.FileName, newImage.FileName, newImage.Content);
+                db.AddFile(newBackground.FileName, newBackground.FileName, newBackground.Content);
+                games[0].Icon = newIcon.FileName;
+                games[0].Image = newImage.FileName;
+                games[0].BackgroundImage = newBackground.FileName;
                 db.UpdateGameInDatabase(games[0]);
 
                 Assert.AreEqual(6, db.Database.FileStorage.FindAll().ToList().Count());
 
                 games = db.GamesCollection.FindAll().ToList();
                 model = new GameEditViewModel(games, db, new MockWindowFactory(), new MockDialogsFactory(), new MockResourceProvider());
-                model.EditingGame.Icon = newIcon.Path;
-                model.EditingGame.Image = newImage.Path;
-                model.EditingGame.BackgroundImage = newBackground.Path;
+                model.EditingGame.Icon = newIcon.FileId;
+                model.EditingGame.Image = newImage.FileId;
+                model.EditingGame.BackgroundImage = newBackground.FileId;
                 model.ConfirmDialog();
 
                 Assert.AreEqual(3, db.Database.FileStorage.FindAll().ToList().Count());
@@ -192,9 +192,9 @@ namespace PlayniteUITests.ViewModels
                 games = db.GamesCollection.FindAll().ToList();
                 foreach (var game in games)
                 {
-                    Assert.AreEqual(newIcon.Name, game.Icon);
-                    Assert.AreEqual(newImage.Name, game.Image);
-                    Assert.AreEqual(newBackground.Name, game.BackgroundImage);
+                    Assert.AreEqual(newIcon.FileName, game.Icon);
+                    Assert.AreEqual(newImage.FileName, game.Image);
+                    Assert.AreEqual(newBackground.FileName, game.BackgroundImage);
                 }
             }
         }
